src/cost_export/requirements.txt
azure-functions
pyarrow
pandas
azure-identity
boto3
s3fs
azure-storage-blob